# Refactored and optimized rendering